fix: resolve all Hifz compilation errors
- Add missing Hifz models (HifzGoalModel, StudentXPModel, VerseProgressModel, etc.)
- Add Hifz API methods to LearningApiService
- Fix provider imports in all hifz screens (use hifz_provider.dart)
- Fix router: add moduleNumber params, surahName param
- Fix Icons.close_circle → Icons.cancel, type cast issues

+++ lib/features/autonomous_learning/screens/learning_hub_screen.dart
@@ -34,7 +34,7 @@ class LearningHubScreen extends ConsumerWidget {
           final avgMastery = moduleProgress.isEmpty
               ? 0
               : moduleProgress.values
-                      .fold<int>(0, (sum, m) => sum + (m['masteryLevel'] ?? 0)) ~/
+                      .fold<int>(0, (sum, m) => sum + ((m['masteryLevel'] ?? 0) as int)) ~/
                   moduleProgress.length;
           final comprehensionScore = (avgMastery / 3 * 100).toInt();
